requirements.txt
fastapi
uvicorn
python-dotenv
pandas
openpyxl
langchain
langchain-community
langchain-core
ollama
faiss-cpu
langchain-ollama
chromadb


